Restructure code. Implement full RunProperties
