+++ src/elements/document.rs
@@ -1,4 +1,4 @@
-use crate::elements::paragraph::Paragraph;
+use crate::elements::Paragraph;
 #[derive(Debug, Default, Clone, PartialEq)]
 pub struct Document {
     pub paragraphs: Vec<Paragraph>,
