+++ src/elements/run.rs
@@ -1,4 +1,4 @@
-use crate::elements::run_properties::RunProperties;
+use crate::elements::RunProperties;
 #[derive(Debug, Clone, PartialEq)]
 pub struct Run {
     pub properties: RunProperties,
